Imputation (mice)

The main function of the package is mice, which takes a DataFrame as its input. It returns a multiply imputed dataset (Mids) object with the imputed values.

Mice.Mids — Type

Mids

A multiply imputed dataset object.

The data originally supplied are stored as data.

The imputed data are stored as imputations (one column per imputation).

The number of imputations is stored as m.

The imputation method for each variable is stored as methods.

The predictor matrix is stored as predictorMatrix.

The order in which the variables are imputed is stored as visitSequence.

The number of iterations is stored as iter.

The mean of each variable across the imputations is stored as meanTraces.

The variance of each variable across the imputations is stored as varTraces.

Mice.mice — Function

mice(
    data::DataFrame;
    m::Int = 5,
    visitSequence::Union{Vector{String}, Nothing} = nothing,
    methods::Union{NamedVector{String}, Nothing} = nothing,
    predictorMatrix::Union{NamedMatrix{Bool}, Nothing} = nothing,
    iter::Int = 10,
    progressReports::Bool = true,
    gcSchedule::Float64 = 1.0,
    threads::Bool = true,
    kwargs...
    )

Imputes missing values in a dataset using the MICE algorithm. Heavily based on the R package mice (van Buuren & Groothuis-Oudshoorn, 2011).

The data containing missing values (data) must be supplied as a DataFrame.

The number of imputations created is specified by m.

The variables will be imputed in the order specified by visitSequence. The default is sorted by proportion of missing data in ascending order; the order can be customised using a vector of variable names in the desired order.

The imputation method for each variable is specified by the NamedArray methods. The default is to use predictive mean matching (pmm) for all variables. Currently only pmm is supported. Any variable not to be imputed can be marked as such using an empty string ("").

The predictor matrix is specified by the NamedArray predictorMatrix. The default is to use all other variables as predictors for each variable. Any variable not predicting another variable can be marked as such in the matrix using a 0.

The number of iterations is specified by iter.

If progressReports is true, a progress indicator will be displayed in the console.

gcSchedule dictates when the garbage collector will be (additionally) invoked. The number provided is the fraction of your RAM remaining at which the GC will be called. For small datasets, you may get away with a value of 0.0 (never called), but for larger datasets, it may be worthwhile to call it more frequently. The default is to call it after each iteration of each variable (1.0), but this may negatively affect performance if it is not necessary for your dataset.

threads dictates whether multi-threading will be used. This will improve performance for larger jobs if and only if Julia has been launched with multiple threads (which you can verify by calling Threads.nthreads()). The default is true.

mice(
    mids::Mids;
    iter::Int = 10,
    progressReports::Bool = true,
    gcSchedule::Float64 = 1.0,
    threads::Bool = true,
    kwargs...
    )

Adds additional iterations to an existing Mids object.

The number of additional iterations is specified by iter.

If progressReports is true, a progress indicator will be displayed in the console.

gcSchedule dictates when the garbage collector will be (additionally) invoked. The number provided is the fraction of your RAM remaining at which the GC will be called. For small datasets, you may get away with a value of 0.0 (never called), but for larger datasets, it may be worthwhile to call it more frequently. The default is to call it after each iteration of each variable (1.0), but this may negatively affect performance if it is not necessary for your dataset.

threads dictates whether multi-threading will be used. This will improve performance for larger jobs if and only if Julia has been launched with multiple threads (which you can verify by calling Threads.nthreads()). The default is true.

Mice.makeMonotoneSequence — Function

makeMonotoneSequence(data::DataFrame)

Returns a vector of the column names in a DataFrame in ascending order of missingness. This facilitates convergence in cases where missingness follows a "monotone" pattern. It is the default visit sequence for the mice() function.

Mice.makePredictorMatrix — Function

makePredictorMatrix(data::DataFrame)

Returns a named matrix of booleans defining the predictors for each variable in data. The variables to be predicted are on the rows, and the predictors are on the columns. The default is to use all variables as predictors for all other variables (i.e. all 1s except for the diagonal, which is 0).

Mice.makeMethods — Function

makeMethods(data::DataFrame)

Returns a named vector of strings defining the method by which each variable in data should be imputed in the mice() function. The default (and only supported) method is predictive mean matching (pmm).

Diagnostics

After performing multiple imputation, you should inspect the trace plots of the imputed variables to verify convergence. Mice.jl includes the a plotting function to do this.

RecipesBase.plot — Function

plot(
    mids::Mids,
    var::String
    )

Plots the mean and standard deviation of the imputed values for a given variable. Here var is given as a string (the name of the variable).

plot(
    mids::Mids,
    var_no::Int
    )

Plots the mean and standard deviation of the imputed values for a given variable. Here var_no is given as an integer (the index of the variable in the visitSequence).